requirements.txt
Django==3.1.3
Pillow==8.0.1
psycopg2-binary==2.8.6
django-crispy-forms==1.9.2
python-dateutil==2.8.1
